+++ src/main/java/com/pro/infomate/email/repository/EmailAndMemberRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface EmailAndMemberRepository extends JpaRepository<Email, Long> {
+public interface EmailAndMemberRepository extends JpaRepository<Email, Integer> {
     @Query("SELECT c FROM Email c join c.member m where m.memberCode = :memberCode")
-    List<Email> findByMemberCode(Long memberCode);
+    List<Email> findByMemberCode(Integer memberCode);
 }
